fix: Replace problematic manual review migration with idempotent version
- Created new add_manual_review_tables_v2 migration
- Added checks for existing tables, columns, and enum types
- Made migration fully idempotent to handle partial runs
- Removed old problematic migration that was causing enum conflicts
- This resolves the "type reviewreason already exists" error
